# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckRequest": {
        "properties": {
          "variant": {
            "default": "cpu",
            "title": "Variant",
            "type": "string"
          }
        },
        "title": "CheckRequest",
        "type": "object"
      },
...
      "JobDoc": {
        "description": "A queue job: what to do, where it stands, and what came of it.",
        "properties": {
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "lane": {
            "title": "Lane",
            "type": "string"
          },
          "owner": {
            "default": "local",
            "title": "Owner",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "worker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Id"
          }
        },
        "required": [
          "id",
          "kind",
          "lane",
          "state",
          "payload",
          "created_at"
        ],
        "title": "JobDoc",
        "type": "object"
      },
      "LessonDetail": {
        "properties": {
          "checks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Checks",
            "type": "array"
          },
          "compute": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ComputeDoc"
            },
            "title": "Compute",
            "type": "object"
          },
          "depth": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Depth",
            "type": "object"
          },
          "experiment": {
            "additionalProperties": true,
            "title": "Experiment",
            "type": "object"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "forbid": {
            "items": {
              "type": "string"
            },
            "title": "Forbid",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "integer"
          },
          "locked_by": {
            "items": {
              "type": "string"
            },
            "title": "Locked By",
            "type": "array"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisites",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "text": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Text",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "unlocks": {
            "items": {
              "type": "string"
            },
            "title": "Unlocks",
            "type": "array"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "summary",
          "level",
          "state",
          "locked_by",
          "unlocks",
          "compute",
          "prerequisites",
          "forbid",
          "checks",
          "experiment",
          "depth",
          "text",
          "files",
          "workspace"
        ],
        "title": "LessonDetail",
        "type": "object"
      },
...
      "PathSummary": {
        "properties": {
          "compute": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ComputeDoc"
            },
            "title": "Compute",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonSummary"
            },
            "title": "Lessons",
            "type": "array"
          },
          "level": {
            "title": "Level",
            "type": "integer"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisites",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "level",
          "summary",
          "prerequisites",
          "compute",
          "lessons"
        ],
        "title": "PathSummary",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": true,
        "description": "The body of every error response (RFC 9457).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "problems": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problems"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "nanoscope",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/curricula": {
      "get": {
        "description": "Every learning path with its lessons, state and compute estimates.",
        "operationId": "curricula_api_curricula_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PathSummary"
                  },
                  "title": "Response Curricula Api Curricula Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Curricula",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/curricula/{path}/{lesson}": {
      "get": {
        "description": "One lesson: its text (Surface, Deep, Reading), checks, unlocks and starter files.",
        "operationId": "lesson_detail_api_curricula__path___lesson__get",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lesson",
            "required": true,
            "schema": {
              "title": "Lesson",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Lesson Detail",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/curricula/{path}/{lesson}/check": {
      "post": {
        "description": "Run the lesson's checks on your file as a job (they train models and run your code, so\na worker does it). The result lands in `learn/checks/` and in `/learn/progress`.",
        "operationId": "check_api_curricula__path___lesson__check_post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lesson",
            "required": true,
            "schema": {
              "title": "Lesson",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CheckRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDoc"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Check",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/curricula/{path}/{lesson}/predict": {
      "post": {
        "description": "Commit to a prediction before the experiment runs (what `nanoscope learn predict` does).\nOnce the lesson has been checked it is too late: a 422 says so.",
        "operationId": "predict_api_curricula__path___lesson__predict_post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lesson",
            "required": true,
            "schema": {
              "title": "Lesson",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PredictRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Predict",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/curricula/{path}/{lesson}/start": {
      "post": {
        "description": "Copy the lesson's starter files into your workspace (never over your edits) and mark it\nstarted. The very first start turns gating on (`guided`) unless you ask for `open`.",
        "operationId": "start_api_curricula__path___lesson__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lesson",
            "required": true,
            "schema": {
              "title": "Lesson",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StartRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Start",
        "tags": [
          "learn"
        ]
      }
    },
...
    "/api/learn/events": {
      "get": {
        "description": "Server-sent events: `unlocks` or `progress` with the new document whenever either file\nchanges (a lesson check finished, `nanoscope learn unlock` ran in a terminal).",
        "operationId": "learn_events_api_learn_events_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Learn Events",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/learn/policy": {
      "post": {
        "description": "`open` unlocks everything; `guided` turns gating back on, keeping what you earned or\nskipped.",
        "operationId": "learn_policy_api_learn_policy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnlocksView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Learn Policy",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/learn/progress": {
      "get": {
        "description": "Where you are in each lesson (progress.json).",
        "operationId": "learn_progress_api_learn_progress_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressDoc"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Learn Progress",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/learn/unlock": {
      "post": {
        "description": "Unlock everything (`all`), or skip one lesson's lock with a reason. Both are recorded, so\nyour progress shows what you earned and what you skipped.",
        "operationId": "learn_unlock_api_learn_unlock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnlocksView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Learn Unlock",
        "tags": [
          "learn"
        ]
      }
    },
    "/api/learn/unlocks": {
      "get": {
        "description": "The gating policy, what you earned or skipped, and what is still locked.",
        "operationId": "learn_unlocks_api_learn_unlocks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnlocksView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Learn Unlocks",
        "tags": [
          "learn"
        ]
      }
    },
...
  }
}